fix headers/header guards
