+++ include/ti_mmwave_rospkg/DataHandlerClass.h
@@ -1,21 +1,61 @@
-#ifndef _DATA_HANDLER_CLASS_
-#define _DATA_HANDLER_CLASS_
+/*
+ * DataHandlerClass.h
+ *
+ * This file contains various defines used within this package.
+ *
+ *
+ * Copyright (C) 2017 Texas Instruments Incorporated - http://www.ti.com/ 
+ *
+ *
+ *  Redistribution and use in source and binary forms, with or without 
+ *  modification, are permitted provided that the following conditions 
+ *  are met:
+ *
+ *    Redistributions of source code must retain the above copyright 
+ *    notice, this list of conditions and the following disclaimer.
+ *
+ *    Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the 
+ *    documentation and/or other materials provided with the   
+ *    distribution.
+ *
+ *    Neither the name of Texas Instruments Incorporated nor the names of
+ *    its contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS 
+ *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT 
+ *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ *  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT 
+ *  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, 
+ *  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT 
+ *  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ *  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ *  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT 
+ *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE 
+ *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+*/
 
-#include <ti_mmwave_rospkg/RadarScan.h>
-#include "mmWave.h"
-#include "point_types.h"
+
+#ifndef TI_MMWAVE_ROSPKG_DATAHANDLERCLASS_H
+#define TI_MMWAVE_ROSPKG_DATAHANDLERCLASS_H
+
+#include "ti_mmwave_rospkg/RadarScan.h"
+#include "ti_mmwave_rospkg/mmWave.h"
+#include "ti_mmwave_rospkg/point_types.h"
+
 #include <iostream>
 #include <cstdio>
 #include <cstdlib>
 #include <boost/shared_ptr.hpp>
-#include "ros/ros.h"
-#include "sensor_msgs/PointCloud2.h"
+#include <ros/ros.h>
 #include <pthread.h>
 #include <algorithm>
-#include "pcl_ros/point_cloud.h"
-#include "sensor_msgs/PointField.h"
-#include "sensor_msgs/PointCloud2.h"
-#include "sensor_msgs/point_cloud2_iterator.h"
+#include <pcl_ros/point_cloud.h>
+#include <sensor_msgs/PointField.h>
+#include <sensor_msgs/PointCloud2.h>
+#include <sensor_msgs/point_cloud2_iterator.h>
 #include <pcl/io/pcd_io.h>
 #include <pcl/point_types.h>
 #include <visualization_msgs/Marker.h>
@@ -137,4 +177,4 @@ private:
     ros::Publisher DataUARTHandler_pub;
 };
 
-#endif
+#endif  // TI_MMWAVE_ROSPKG_DATAHANDLERCLASS_H
