+++ include/ti_mmwave_rospkg/ParameterParser.h
@@ -1,35 +1,72 @@
-#ifndef _PARAM_PARSER_CLASS_
-#define _PARAM_PARSER_CLASS_
+/*
+ * ParameterParser.h
+ *
+ * ParameterParser description...
+ *
+ *
+ * Copyright (C) 2017 Texas Instruments Incorporated - http://www.ti.com/
+ *
+ *
+ *  Redistribution and use in source and binary forms, with or without
+ *  modification, are permitted provided that the following conditions
+ *  are met:
+ *
+ *    Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *
+ *    Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in the
+ *    documentation and/or other materials provided with the
+ *    distribution.
+ *
+ *    Neither the name of Texas Instruments Incorporated nor the names of
+ *    its contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ *  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ *  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ *  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ *  A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ *  OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ *  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ *  LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ *  DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ *  THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ *  (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ *  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ */
 
-#include "ros/ros.h"
-#include "std_msgs/String.h"
-#include "ti_mmwave_rospkg/mmWaveCLI.h"
+#ifndef TI_MMWAVE_ROSPKG_PARAMETERPARSER_H
+#define TI_MMWAVE_ROSPKG_PARAMETERPARSER_H
+
 #include <nodelet/nodelet.h>
 #include <pluginlib/class_list_macros.h>
+#include <ros/ros.h>
+#include <std_msgs/String.h>
+
+#include <cstdio>
 #include <cstdlib>
 #include <fstream>
 #include <iostream>
-#include <cstdio>
 #include <sstream>
 #include <string>
 #include <vector>
 
-namespace ti_mmwave_rospkg {
+#include "ti_mmwave_rospkg/mmWaveCLI.h"
 
-class ParameterParser : public nodelet::Nodelet{
+namespace ti_mmwave_rospkg
+{
+class ParameterParser : public nodelet::Nodelet
+{
+public:
+  ParameterParser();
+  void ParamsParser(ti_mmwave_rospkg::mmWaveCLI &srv, ros::NodeHandle &n);
+  void CalParams(ros::NodeHandle &nh);
 
-  public:
-  	
-  	ParameterParser();
-  	void ParamsParser(ti_mmwave_rospkg::mmWaveCLI &srv, ros::NodeHandle &n);
-	void CalParams(ros::NodeHandle &nh);
-
-  private:
-    
-    virtual void onInit();
-    
-    ti_mmwave_rospkg::mmWaveCLI srv;
-
+private:
+  virtual void onInit();
+  ti_mmwave_rospkg::mmWaveCLI srv;
 };
-}
-#endif
+}  // namespace ti_mmwave_rospkg
+#endif  // TI_MMWAVE_ROSPKG_PARAMETERPARSER_H
